big reorganization of builtins folder. Fix #56. Bug in env var with _ fixed too.
